raspberrypi_dev/img_detect/dev/requirements.txt
numpy
opencv-python
RPi.GPIO
pyserial
pillow